refactor: fix file structure
