+++ FE/vite.config.js
@@ -7,4 +7,7 @@ import react from "@vitejs/plugin-react";
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  build: {
+    outDir: "dist", 
+  },
 });
